# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive API documentation for the Ecommerce Platform. All endpoints are accessed through the API Gateway which acts as a reverse proxy to backend microservices (Identity Service, Product Service, Order Service, etc.). The Gateway automatically forwards cookies, headers, and user context to backend services.",
        "title": "Ecommerce Platform API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/cart": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Lấy danh sách sản phẩm trong giỏ hàng của người dùng hiện tại (yêu cầu đăng nhập)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Lấy nội dung giỏ hàng",
                "responses": {
                    "200": {
                        "description": "Thông tin giỏ hàng",
                        "schema": {
                            "$ref": "#/definitions/models.CartResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Lỗi hệ thống",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Xóa tất cả sản phẩm trong giỏ hàng của người dùng",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Xóa toàn bộ giỏ hàng",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Lỗi hệ thống",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cart/items": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Thêm một sản phẩm cụ thể (SKU) vào giỏ hàng với số lượng nhất định",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Thêm sản phẩm vào giỏ",
                "parameters": [
                    {
                        "description": "Thông tin sản phẩm thêm vào",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddToCartRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CartResponse"
                        }
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ hoặc thiếu hàng",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cart/items/{product_item_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Thay đổi số lượng của một SKU cụ thể trong giỏ hàng",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Cập nhật số lượng sản phẩm",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của SKU sản phẩm",
                        "name": "product_item_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Số lượng mới",
                        "name": "quantity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateCartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CartResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Xóa một sản phẩm cụ thể ra khỏi giỏ hàng",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Xóa sản phẩm khỏi giỏ hàng",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của SKU sản phẩm",
                        "name": "product_item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Tham số không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy sản phẩm",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Lỗi hệ thống",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cart/items/{product_item_id}/selection": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set the selection state (selected/deselected) for a cart item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Set item selection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của SKU sản phẩm",
                        "name": "product_item_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Trạng thái chọn",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SelectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy sản phẩm",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Lỗi hệ thống",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cart/items/{product_item_id}/toggle": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Thay đổi trạng thái được chọn (tick) của một sản phẩm để chuẩn bị thanh toán",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Bật/Tắt chọn sản phẩm",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của SKU sản phẩm",
                        "name": "product_item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CartResponse"
                        }
                    }
                }
            }
        },
        "/cart/select_all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Chọn hoặc bỏ chọn toàn bộ sản phẩm có trong giỏ hàng",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Chọn/Bỏ chọn tất cả",
                "parameters": [
                    {
                        "description": "Trạng thái chọn (true/false)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SelectAllRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/cart/selected": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Xóa tất cả sản phẩm đang được tick chọn ra khỏi giỏ hàng",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Xóa các mục đã chọn",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/cart/selection": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set selection state for all items in the cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Set selection for all items",
                "parameters": [
                    {
                        "description": "Trạng thái chọn",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SelectAllRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Lỗi hệ thống",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cart/shops/{shop_id}/select": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Chọn hoặc bỏ chọn tất cả sản phẩm thuộc về một cửa hàng cụ thể",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Chọn sản phẩm theo Shop",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của cửa hàng",
                        "name": "shop_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Trạng thái chọn",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SelectAllRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/cart/shops/{shop_id}/selection": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set selection state for all items from a shop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Set selection for shop items",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của cửa hàng",
                        "name": "shop_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Trạng thái chọn",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SelectShopRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Lỗi hệ thống",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cart/validate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Kiểm tra tồn kho và trạng thái của các sản phẩm đã chọn trước khi chuyển sang bước thanh toán",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Kiểm tra giỏ hàng trước thanh toán",
                "responses": {
                    "200": {
                        "description": "Kết quả kiểm tra",
                        "schema": {
                            "$ref": "#/definitions/models.CartValidationResponse"
                        }
                    },
                    "400": {
                        "description": "Sản phẩm không hợp lệ hoặc hết hàng",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddToCartRequest": {
            "type": "object",
            "properties": {
                "product_item_id": {
                    "type": "integer",
                    "example": 1001
                },
                "quantity": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.CartResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CartItem"
                    }
                },
                "total_items": {
                    "type": "integer",
                    "example": 2
                },
                "total_price": {
                    "type": "number",
                    "example": 30200
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-01-15T12:00:00Z"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Bad request"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        },
...
        "models.SelectAllRequest": {
            "type": "object",
            "properties": {
                "selected": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SelectShopRequest": {
            "type": "object",
            "properties": {
                "selected": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SelectionRequest": {
            "type": "object",
            "properties": {
                "selected": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                }
            }
        },
...
        "models.UpdateCartRequest": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT access token. Example: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "HttpOnly cookie containing JWT access token. Automatically set after login/register. Used for authentication in browser-based applications.",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}